Removes hard-coded file names from NuSpec file.
The NuSpec file had a number of files hard-coded in it which made it brittle when it comes to changing filesets.  By utilizing wild cards, we can reliably change the included set of files by simply adding files to the repo rather than by adding them to the repo and also updating the nuspec file.